initial web app refactoring (separated py and html, js to static)
