+++ server_side/libraries/libraries.py
@@ -12,6 +12,7 @@ import threading
 import time
 from pymongo import MongoClient
 import pymongo
+from jinja2 import Environment, FileSystemLoader
 
 class UrlLibThread(threading.Thread):
     def __init__(self, books_ids, book_metadata_url, domain, tunnel, base_url):
@@ -127,8 +128,8 @@ class JSONBooks:
                 self.books_ids = requests.get("{base_url}{books_ids_url}".format(base_url=self.base_url, books_ids_url=self.books_ids_url)).json()['book_ids']
                 if not self.books_ids:
                     continue
-                except:
-                    continue
+            except:
+                continue
 
             self.book_metadata_url = 'ajax/book/'
             thrd = UrlLibThread(self.books_ids, self.book_metadata_url, self.domain, self.tunnel, self.base_url)
@@ -180,179 +181,16 @@ class Root(object):
 
     @cherrypy.expose
     def index(self):
-        return """
-<html>
-<head><title>Libraries</title>
-<script type="application/javascript" src="static/jquery-1.9.1.js"></script>
-<script type="application/javascript" src="static/jquery-ui-1.10.3.custom.min.js"></script>
-<link rel="stylesheet" type="text/css" href="static/style.css" />
-<link rel="stylesheet" type="text/css" href="static/jquery-ui-1.10.3.custom.min.css" />
-<script type='text/javascript' charset='utf-8'>
-var LSB = {}
-LSB.start = 0;
-LSB.total_num = 1;
-LSB.offset = 8;
-LSB.query = "";
-LSB.carry = "";
-LSB.processing = "";
-LSB.prefix_url = "http://www"
+        tmpl = ENV.get_template('index.html')
+        return tmpl.render()
 
-
-init_page = function() {
-                add_toolbar();
-                render_page();
-}
-
-render_page = function() {
-
-    $.ajax({
-      type: 'POST',
-      url: "render_page",
-      contentType: "application/json",
-      processData: false,
-      data: JSON.stringify(LSB),
-      success: function(books) {
-                    $('#content').empty();
-                    window.foobar = books;
-                    add_toolbar();
-                    toolbar_data = books.pop()
-                    LSB.total_num = toolbar_data['total_num']
-                    LSB.query = toolbar_data['query']
-                    LSB.processing = toolbar_data['processing']
-                    refresh_pagination()
-                    $(function () {
-                            $('#authors').autocomplete({source: toolbar_data['authors'], minLength:2});
-                            $('#titles').autocomplete({source: toolbar_data['titles'], minLength:2});
-                        });
-                    $.each(books, function(n, book) {
-                        window.barfoo = book;
-                        var base_url = LSB.prefix_url + book.tunnel + '.' + book.domain
-                        var formats = ""
-                        var authors = '<div id="authorz">'
-
-                        book.formats.map(function(format) { 
-                            formats = formats + '<a href="' + base_url + '/get/' + format + '/' + book.id +'.' + format + '">' + format.toUpperCase() + '</a> '});
-
-                         book.authors.map(function(author) {
-                                author_s = author.replace("'", " ")
-                                author_param = 'search_author("' + author_s + '")'
-                                authors = authors + "<a id='author' href='#' title='show only books by " + author_s  + "' onClick='" + author_param + "'>" + author + ", </a>&nbsp;"});
-
-                        last_comma = authors.lastIndexOf(",");
-                        authors = authors.substr(0, last_comma) + authors.substr(last_comma + 1) + '</div>'
-                        $('#content').append('<div class="cover"><a href="'+ base_url +'/browse/book/'+ book.id +'" target="_blank" id="more_about" title="about this book"><img src="' + base_url + '/get/cover/' + book.id + '.jpg"></img></a><h2><a href="' + base_url + '/browse/book/' + book.id + '" target="_blank" id="more_about" title="about this book">' + book.title + '</a><br/>' + authors  + '</h2><span class="download">Metadata: <a href="'+ base_url + '/get/opf/' + book.id  + book.title.replace(/\?/g, "") + '.opf" title="import metadata directly to calibre">.opf</a><br/>Download: ' + formats + ' </span></div>')
-                });
-
-    },
-      dataType: "json"
-    });
-}
-
-refresh_pagination = function () {
-    //$('.pagination').button({label: ' ' + (+LSB.start + 1)+ ' - ' + (+LSB.start + +LSB.offset) + ' out of ' + LSB.total_num + ' books ' + LSB.processing, disabled: true})
-    $('.pagination').button({label: 'HOME (' + (+LSB.start + 1)+ ' - ' + (+LSB.start + +LSB.offset) + ' out of ' + LSB.total_num + ' books)' + LSB.processing}).click(function() {search_query()})
-}
-
-add_toolbar = function() {
-    $('#content').append('<div id="toolbar"><div id="prev_page"></div><div id="next_page"></div></div>');
-    $('#content').append('<div id="searchbar"></div>')
-
-    $('#toolbar').append($('#prev_page').button({label: '<<<'}));
-    $('#prev_page').click(function() {prev_page()});
-    $('#toolbar').append($('#next_page').button({label: '>>>'}));
-    $('#next_page').click(function() {next_page()});
-    $('#toolbar').append('<div class="pagination"></div>');
-
-    $('#searchbar').append('<div class="ui-widget"><input id="authors" placeholder="authors"/><input id="titles" placeholder="titles"/><input id="search_all" placeholder="search all metadata"/><div id="search"></div></div>')
-    $('#search').button({label: 'SEARCH'}).click(search_query)
-
-}
-
-search_query = function () {
-    if ($('#authors').val()) {
-        LSB.query = 'authors:' + $('#authors').val()
-        LSB.carry = " OR "
-    } else {
-        LSB.query = '';
-        LSB.carry = '';
-    }
-    if ($('#titles').val()) {
-        LSB.query += LSB.carry + 'title:' + $('#titles').val()
-        LSB.carry = " OR "
-    } else if (LSB.carry != " OR ") {
-        LSB.query = '';
-        LSB.carry = '';
-    }
-
-    if ($('#search_all').val()) {
-        LSB.query += LSB.carry + $('#search_all').val()
-    } else if (LSB.carry != " OR ") {
-        LSB.query = '';
-        LSB.carry = '';
-    }
-    
-    LSB.start = 0;
-    render_page()
-    LSB.query = '';
-    LSB.carry = '';
-}
-
-search_author = function(author) {
-        LSB.query = "authors:" + author;
-        console.log(LSB.query);
-        LSB.start = 0;
-        render_page();
-        LSB.query = "";
-}
-
-next_page = function() {
-    LSB.start = LSB.start + LSB.offset;
-    if (LSB.start+LSB.offset >= LSB.total_num) {
-        LSB.start = LSB.total_num - LSB.offset;
-    }
-    render_page();
-}
-
-prev_page = function() {
-    LSB.start = LSB.start - LSB.offset;
-    if (LSB.start <= 0) {
-        LSB.start = 0;
-    }
-    render_page();
-}
-
-$(document).ajaxStart(function() { 
-        $('body').addClass("loading"); 
-    });
-
-$(document).ajaxStop(function() { 
-        $('body').removeClass("loading"); 
-    });
-
-$(function() {
-    $(document).tooltip({track:true});
-  });
-
-$(document).ready(function() {
-    init_page();
-});
-
-</script>
-</title>
-<body>
-<div id="content">
-</div>
-<div class="modal"></div>
-</body>
-</html>
-"""
-
+ENV = Environment(loader=FileSystemLoader('templates'))
 base_dir = "/var/www/libraries/"
 prefix_url = "http://www"
 current_dir = os.path.dirname(os.path.abspath(__file__))
 conf = {'/static': {'tools.staticdir.on': True,
                     'tools.staticdir.dir': os.path.join(current_dir, 'static'),
-                    'tools.staticdir.content_types': {'javascript': 'application/javascript',
+                    'tools.staticdir.content_types': {'js': 'application/javascript',
                                                       'css': 'text/css',
                                                       'gif': 'image/gif'
                                                        }}}
