+++ server_side/libraries/templates/index.html
@@ -1,0 +1,16 @@
+<html>
+  <head>
+    <title>Libraries</title>
+    <script type="application/javascript" src="static/jquery-1.9.1.js"></script>
+    <script type="application/javascript" src="static/jquery-ui-1.10.3.custom.min.js"></script>
+    <script type="application/javascript" src="static/libraries.js"></script>
+    <link rel="stylesheet" type="text/css" href="static/style.css" />
+    <link rel="stylesheet" type="text/css" href="static/jquery-ui-1.10.3.custom.min.css" />
+  </head>
+  <body>
+    test
+    <div id="content">
+    </div>
+    <div class="modal"></div>
+  </body>
+</html>
